Refactor image recommendations onboarding into AltTextImageRecommendationsOnboardingPresenter
